+++ resolwe/elastic/indices.py
@@ -360,6 +360,8 @@ class BaseIndex(object):
         index_name = self.document_class()._get_index()  # pylint: disable=protected-access,not-callable
         connections.get_connection().indices.delete(index_name, ignore=404)  # pylint: disable=no-member
 
+        self._mapping_created = False
+
     def get_permissions(self, obj):
         """Return users and groups with ``view`` permission on the current object.